curametric_backend/curametric_wound_api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from .views import UserViewSet, PatientViewSet, WoundViewSet, WoundCareViewSet
from .views import GoogleLoginView

router = DefaultRouter()
router.register(r'users', UserViewSet)
router.register(r'patients', PatientViewSet)
router.register(r'wounds', WoundViewSet)
router.register(r'woundcares', WoundCareViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path("google-login/", GoogleLoginView.as_view(), name="google_login"),]
