curametric_backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environ==1.0
google-auth==2.38.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
